% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layouttree.R
\name{layouttree}
\alias{layouttree}
\title{Utility function}
\usage{
layouttree(Tree)
}
\arguments{
\item{Tree}{an object of the class "ranktree"}
}
\value{
an object of the class "ranktree" completing the output of the function \code{ranktree}
}
\description{
A utility function completing the output of the function \code{ranktree}.
}
\author{
Antonio D'Ambrosio \email{antdambr@unina.it}
}
