% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.cca.R
\name{print.cca}
\alias{print.cca}
\title{S3 methods for cca}
\usage{
\method{print}{cca}(x, ...)
}
\arguments{
\item{x}{An object of the class "cca"}

\item{\dots}{not used}
}
\value{
print a brief summary of the CCA
}
\description{
Print methods for objects of class \code{cca}
}
