% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.ranktree.R
\name{print.ranktree}
\alias{print.ranktree}
\title{S3 methods for ranktree}
\usage{
\method{print}{ranktree}(x, ...)
}
\arguments{
\item{x}{An object of the class "ranktree"}

\item{\dots}{not used}
}
\value{
print a brief summary of the prediction tree
}
\description{
Print methods for objects of class \code{ranktree}
}
\examples{

data("Univranks")
tree <- ranktree(Univranks$rankings,Univranks$predictors,num=50)
tree



}
