% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.cca.R
\name{summary.cca}
\alias{summary.cca}
\title{S3 methods for ranktree}
\usage{
\method{summary}{cca}(object, ...)
}
\arguments{
\item{object}{An object of the class "cca"}

\item{\dots}{not used}
}
\value{
it shows the summary of the prediction tree
}
\description{
Summary methods for objects of class \code{cca}
}
