% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.ranktree.R
\name{summary.ranktree}
\alias{summary.ranktree}
\title{S3 methods for ranktree}
\usage{
\method{summary}{ranktree}(object, ...)
}
\arguments{
\item{object}{An object of the class "ranktree"}

\item{\dots}{not used}
}
\value{
it shows the summary of the prediction tree
}
\description{
Summary methods for objects of class \code{ranktree}
}
\examples{

data("Univranks")
tree <- ranktree(Univranks$rankings,Univranks$predictors,num=50)
summary(tree)

}
